[Budget] changed arrays to list in Budget class
