+++ Budget/src/main/java/lt/codeacademy/BudgetProgram.java
@@ -1,3 +1,5 @@
+package lt.codeacademy;
+
 public class BudgetProgram {
   public static void main(String[] args) {
     BudgetRun.run();
