+++ Budget/src/main/java/lt/codeacademy/IncomeRecord.java
@@ -1,7 +1,10 @@
+package lt.codeacademy;
+
 import java.time.LocalDateTime;
 
-public class ExpensesRecord extends Record {
-  public ExpensesRecord(
+public class IncomeRecord extends Record {
+
+  public IncomeRecord(
       String amount,
       String date,
       String categoryIndex,
